requirements.txt
qrcode[pil]
pillow
pyzbar
